refactor(ui): move objects/ops related pages into folders
