+++ weave-js/src/components/PagePanelComponents/Home/Browse3.tsx
@@ -72,16 +72,16 @@ import {CompareEvaluationsPage} from './Browse3/pages/CompareEvaluationsPage/Com
 import {LeaderboardListingPage} from './Browse3/pages/LeaderboardPage/LeaderboardListingPage';
 import {LeaderboardPage} from './Browse3/pages/LeaderboardPage/LeaderboardPage';
 import {ModsPage} from './Browse3/pages/ModsPage';
-import {ObjectPage} from './Browse3/pages/ObjectPage';
-import {ObjectVersionPage} from './Browse3/pages/ObjectVersionPage';
+import {ObjectPage} from './Browse3/pages/ObjectsPage/ObjectPage';
+import {ObjectVersionPage} from './Browse3/pages/ObjectsPage/ObjectVersionPage';
 import {
   ObjectVersionsPage,
   WFHighLevelObjectVersionFilter,
-} from './Browse3/pages/ObjectVersionsPage';
-import {OpPage} from './Browse3/pages/OpPage';
-import {OpsPage} from './Browse3/pages/OpsPage';
-import {OpVersionPage} from './Browse3/pages/OpVersionPage';
-import {OpVersionsPage} from './Browse3/pages/OpVersionsPage';
+} from './Browse3/pages/ObjectsPage/ObjectVersionsPage';
+import {OpPage} from './Browse3/pages/OpsPage/OpPage';
+import {OpsPage} from './Browse3/pages/OpsPage/OpsPage';
+import {OpVersionPage} from './Browse3/pages/OpsPage/OpVersionPage';
+import {OpVersionsPage} from './Browse3/pages/OpsPage/OpVersionsPage';
 import {PlaygroundPage} from './Browse3/pages/PlaygroundPage/PlaygroundPage';
 import {ScorersPage} from './Browse3/pages/ScorersPage/ScorersPage';
 import {TablePage} from './Browse3/pages/TablePage';
