+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/OpsPage/OpVersionPage.tsx
@@ -1,28 +1,28 @@
 import {Button} from '@wandb/weave/components/Button';
 import React, {useMemo, useState} from 'react';
 
-import {Icon} from '../../../../Icon';
-import {LoadingDots} from '../../../../LoadingDots';
-import {Tailwind} from '../../../../Tailwind';
-import {useClosePeek} from '../context';
-import {NotFoundPanel} from '../NotFoundPanel';
-import {OpCodeViewer} from '../OpCodeViewer';
-import {DeleteModal, useShowDeleteButton} from './common/DeleteModal';
+import {Icon} from '../../../../../Icon';
+import {LoadingDots} from '../../../../../LoadingDots';
+import {Tailwind} from '../../../../../Tailwind';
+import {useClosePeek} from '../../context';
+import {NotFoundPanel} from '../../NotFoundPanel';
+import {OpCodeViewer} from '../../OpCodeViewer';
+import {DeleteModal, useShowDeleteButton} from '../common/DeleteModal';
 import {
   CallsLink,
   opNiceName,
   OpVersionsLink,
   opVersionText,
-} from './common/Links';
-import {CenteredAnimatedLoader} from './common/Loader';
+} from '../common/Links';
+import {CenteredAnimatedLoader} from '../common/Loader';
 import {
   ScrollableTabContent,
   SimplePageLayoutWithHeader,
-} from './common/SimplePageLayout';
-import {TabUseOp} from './TabUseOp';
-import {useWFHooks} from './wfReactInterface/context';
-import {opVersionKeyToRefUri} from './wfReactInterface/utilities';
-import {OpVersionSchema} from './wfReactInterface/wfDataModelHooksInterface';
+} from '../common/SimplePageLayout';
+import {useWFHooks} from '../wfReactInterface/context';
+import {opVersionKeyToRefUri} from '../wfReactInterface/utilities';
+import {OpVersionSchema} from '../wfReactInterface/wfDataModelHooksInterface';
+import {TabUseOp} from './Tabs/TabUseOp';
 
 export const OpVersionPage: React.FC<{
   entity: string;
