+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/ScorersPage/AnnotationsTab.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import {FilterableObjectVersionsTable} from '../ObjectVersionsPage';
+import {FilterableObjectVersionsTable} from '../ObjectsPage/ObjectVersionsPage';
 
 export const AnnotationsTab: React.FC<{
   entity: string;
